test: ✅ sync enhanced-resolve test cases; migrate to rstest
