+++ napi/tests/exportsField.test.mjs
@@ -1,12 +1,8 @@
-import { describe, it } from "node:test";
+import { describe, it, expect } from "@rstest/core";
 import { ResolverFactory } from "../index.js";
-import * as assert from "node:assert";
 import * as path from "node:path";
-import { fileURLToPath } from "url";
 
-const fixtureDir = fileURLToPath(
-  new URL("../../fixtures/enhanced_resolve/test/fixtures", import.meta.url)
-);
+const fixtureDir = path.resolve("fixtures/enhanced_resolve/test/fixtures");
 
 const fixture = path.resolve(fixtureDir, "exports-field");
 const fixture2 = path.resolve(fixtureDir, "exports-field2");
@@ -32,8 +28,7 @@ describe("exportsFieldPlugin", () => {
 
   it("resolve root using exports field, not a main field", () => {
     const result = resolver.sync(fixture, "exports-field");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture, "node_modules/exports-field/x.js")
     );
   });
@@ -46,8 +41,7 @@ describe("exportsFieldPlugin", () => {
       extensions: [".js"]
     });
     const result = r.sync(fixture, "exports-field/dist/main.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture, "node_modules/exports-field/lib/lib2/main.js")
     );
   });
@@ -59,83 +53,62 @@ describe("exportsFieldPlugin", () => {
       extensions: [".js"]
     });
     const result = r.sync(fixture2, "exports-field/dist/main.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture2, "node_modules/exports-field/lib/browser.js")
     );
   });
 
   it("throw error if extension not provided", () => {
     const result = resolver.sync(fixture2, "exports-field/dist/main");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("should resolve extension without fullySpecified", () => {
     const result = commonjsResolver.sync(fixture2, "exports-field/dist/main");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture2, "node_modules/exports-field/lib/lib2/main.js")
     );
   });
 
   it("resolver should respect condition names", () => {
     const result = resolver.sync(fixture, "exports-field/dist/main.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture, "node_modules/exports-field/lib/lib2/main.js")
     );
   });
 
-  it(
-    "resolver should respect fallback",
-    { skip: "array fallback in exports field directory mappings" },
-    () => {
-      const result = resolver.sync(fixture2, "exports-field/dist/browser.js");
-      assert.strictEqual(
-        result.path,
-        path.resolve(fixture2, "node_modules/exports-field/lib/browser.js")
-      );
-    }
-  );
+  // skip: array fallback in exports field directory mappings
+  it.skip("resolver should respect fallback", () => {
+    const result = resolver.sync(fixture2, "exports-field/dist/browser.js");
+    expect(result.path).toBe(
+      path.resolve(fixture2, "node_modules/exports-field/lib/browser.js")
+    );
+  });
 
-  it(
-    "resolver should respect query parameters #1",
-    { skip: "array fallback in exports field directory mappings" },
-    () => {
-      const result = resolver.sync(
-        fixture2,
-        "exports-field/dist/browser.js?foo"
-      );
-      assert.strictEqual(
-        result.path,
-        path.resolve(fixture2, "node_modules/exports-field/lib/browser.js?foo")
-      );
-    }
-  );
+  // skip: array fallback in exports field directory mappings
+  it.skip("resolver should respect query parameters #1", () => {
+    const result = resolver.sync(fixture2, "exports-field/dist/browser.js?foo");
+    expect(result.path).toBe(
+      path.resolve(fixture2, "node_modules/exports-field/lib/browser.js?foo")
+    );
+  });
 
   it("resolver should respect query parameters #2. Direct matching", () => {
     const result = resolver.sync(fixture2, "exports-field?foo");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
-  it(
-    "resolver should respect fragment parameters #1",
-    { skip: "array fallback in exports field directory mappings" },
-    () => {
-      const result = resolver.sync(
-        fixture2,
-        "exports-field/dist/browser.js#foo"
-      );
-      assert.strictEqual(
-        result.path,
-        path.resolve(fixture2, "node_modules/exports-field/lib/browser.js#foo")
-      );
-    }
-  );
+  // skip: array fallback in exports field directory mappings
+  it.skip("resolver should respect fragment parameters #1", () => {
+    const result = resolver.sync(fixture2, "exports-field/dist/browser.js#foo");
+    expect(result.path).toBe(
+      path.resolve(fixture2, "node_modules/exports-field/lib/browser.js#foo")
+    );
+  });
 
   it("resolver should respect fragment parameters #2. Direct matching", () => {
     const result = resolver.sync(fixture2, "exports-field#foo");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("relative path should work, if relative path as request is used", () => {
@@ -143,8 +116,7 @@ describe("exportsFieldPlugin", () => {
       fixture,
       "./node_modules/exports-field/lib/main.js"
     );
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture, "node_modules/exports-field/lib/main.js")
     );
   });
@@ -154,27 +126,27 @@ describe("exportsFieldPlugin", () => {
       fixture,
       "./node_modules/exports-field/dist/main.js"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("backtracking should not work for request", () => {
     const result = resolver.sync(fixture, "exports-field/dist/../../../a.js");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("backtracking should not work for exports field target", () => {
     const result = resolver.sync(fixture, "exports-field/dist/a.js");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("self-resolving root", () => {
     const result = resolver.sync(fixture, "@exports-field/core");
-    assert.strictEqual(result.path, path.resolve(fixture, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture, "./a.js"));
   });
 
   it("not exported error", () => {
     const result = resolver.sync(fixture, "exports-field/anything/else");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("field name path #1", () => {
@@ -184,8 +156,7 @@ describe("exportsFieldPlugin", () => {
       extensions: [".js"]
     });
     const result = r.sync(fixture3, "exports-field");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture3, "node_modules/exports-field/main.js")
     );
   });
@@ -197,8 +168,7 @@ describe("exportsFieldPlugin", () => {
       extensions: [".js"]
     });
     const result = r.sync(fixture3, "exports-field");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture3, "node_modules/exports-field/main.js")
     );
   });
@@ -210,8 +180,7 @@ describe("exportsFieldPlugin", () => {
       extensions: [".js"]
     });
     const result = r.sync(fixture3, "exports-field");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture3, "node_modules/exports-field/main.js")
     );
   });
@@ -223,8 +192,7 @@ describe("exportsFieldPlugin", () => {
       extensions: [".js"]
     });
     const result = r.sync(fixture2, "exports-field");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture2, "node_modules/exports-field/index.js")
     );
   });
@@ -236,43 +204,41 @@ describe("exportsFieldPlugin", () => {
       extensions: [".js"]
     });
     const result = r.sync(fixture3, "exports-field");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(fixture3, "node_modules/exports-field/index")
     );
   });
 
   it("request ending with slash #1", () => {
     const result = resolver.sync(fixture, "exports-field/");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("request ending with slash #2", () => {
     const result = resolver.sync(fixture, "exports-field/dist/");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("request ending with slash #3", () => {
     const result = resolver.sync(fixture, "exports-field/lib/");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("should throw error if target is invalid", () => {
     const result = resolver.sync(fixture4, "exports-field");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("throw error if exports field is invalid", () => {
     const result = resolver.sync(fixture, "invalid-exports-field");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   // Wildcard pattern tests
   it("should resolve with wildcard pattern #1", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/features/f.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/features/f.js")
     );
   });
@@ -280,8 +246,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #2", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/features/y/y.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/features/y/y.js")
     );
   });
@@ -289,8 +254,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #4", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/features-no-ext/y/y.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/features/y/y.js")
     );
   });
@@ -298,8 +262,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #5", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/middle/nested/f.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/middle/nested/f.js")
     );
   });
@@ -307,8 +270,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #6", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/middle-1/nested/f.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/middle-1/nested/f.js")
     );
   });
@@ -316,8 +278,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #7", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/middle-2/nested/f.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/middle-2/nested/f.js")
     );
   });
@@ -325,8 +286,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #8", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/middle-3/nested/f");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(
         wcFixture,
         "./node_modules/m/src/middle-3/nested/f/nested/f.js"
@@ -337,8 +297,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #9", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/middle-4/f/nested");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/middle-4/f/f.js")
     );
   });
@@ -346,8 +305,7 @@ describe("exportsFieldPlugin", () => {
   it("should resolve with wildcard pattern #10", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/middle-5/f$/$");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(wcFixture, "./node_modules/m/src/middle-5/f$/$.js")
     );
   });
@@ -355,7 +313,7 @@ describe("exportsFieldPlugin", () => {
   it("should throw error if target is 'null'", () => {
     const wcFixture = path.resolve(fixtureDir, "imports-exports-wildcard");
     const result = resolver.sync(wcFixture, "m/features/internal/file.js");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   // extensionAlias with exports field
@@ -371,8 +329,7 @@ describe("exportsFieldPlugin", () => {
       "exports-field-and-extension-alias"
     );
     const result = r.sync(eaFixture, "@org/pkg/string.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(eaFixture, "./node_modules/@org/pkg/dist/string.js")
     );
   });
@@ -389,8 +346,7 @@ describe("exportsFieldPlugin", () => {
       "exports-field-and-extension-alias"
     );
     const result = r.sync(eaFixture, "pkg/string.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(eaFixture, "./node_modules/pkg/dist/string.js")
     );
   });
@@ -407,8 +363,7 @@ describe("exportsFieldPlugin", () => {
       "exports-field-and-extension-alias"
     );
     const result = r.sync(eaFixture, "pkg/string.js");
-    assert.strictEqual(
-      result.path,
+    expect(result.path).toBe(
       path.resolve(eaFixture, "./node_modules/pkg/dist/string.js")
     );
   });
@@ -425,7 +380,7 @@ describe("exportsFieldPlugin", () => {
       "exports-field-and-extension-alias"
     );
     const result = r.sync(eaFixture, "pkg/string.js");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("should throw error with the extensionAlias option #2", () => {
@@ -440,40 +395,27 @@ describe("exportsFieldPlugin", () => {
       "exports-field-and-extension-alias"
     );
     const result = r.sync(eaFixture, "pkg/string.js");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
-  // invalid package target tests (fixture5)
-  it(
-    "invalid package target #1",
-    { skip: "query strings containing ../ treated as invalid targets" },
-    () => {
-      const result = resolver.sync(fixture5, "@exports-field/bad-specifier");
-      assert.strictEqual(
-        result.path,
-        `${path.resolve(fixture5, "./a.js")}?foo=../`
-      );
-    }
-  );
+  // skip: query strings containing ../ treated as invalid targets
+  it.skip("invalid package target #1", () => {
+    const result = resolver.sync(fixture5, "@exports-field/bad-specifier");
+    expect(result.path).toBe(`${path.resolve(fixture5, "./a.js")}?foo=../`);
+  });
 
-  it(
-    "invalid package target #2",
-    { skip: "query strings containing ../ treated as invalid targets" },
-    () => {
-      const result = resolver.sync(
-        fixture5,
-        "@exports-field/bad-specifier/foo/file.js"
-      );
-      assert.strictEqual(
-        result.path,
-        `${path.resolve(fixture5, "./a.js")}?foo=../#../`
-      );
-    }
-  );
+  // skip: query strings containing ../ treated as invalid targets
+  it.skip("invalid package target #2", () => {
+    const result = resolver.sync(
+      fixture5,
+      "@exports-field/bad-specifier/foo/file.js"
+    );
+    expect(result.path).toBe(`${path.resolve(fixture5, "./a.js")}?foo=../#../`);
+  });
 
   it("invalid package target #3", () => {
     const result = resolver.sync(fixture5, "@exports-field/bad-specifier/bar");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #4", () => {
@@ -481,7 +423,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/baz-multi"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #5", () => {
@@ -489,7 +431,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/pattern/a.js"
     );
-    assert.strictEqual(result.path, path.resolve(fixture5, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture5, "./a.js"));
   });
 
   it("invalid package target #6", () => {
@@ -497,7 +439,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/slash"
     );
-    assert.strictEqual(result.path, path.resolve(fixture5, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture5, "./a.js"));
   });
 
   it("invalid package target #7", () => {
@@ -505,7 +447,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/no-slash"
     );
-    assert.strictEqual(result.path, path.resolve(fixture5, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture5, "./a.js"));
   });
 
   it("invalid package target #8", () => {
@@ -513,7 +455,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/utils/index.mjs"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #9", () => {
@@ -521,7 +463,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/utils1/index.mjs"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #10", () => {
@@ -529,7 +471,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/utils2/index"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #11", () => {
@@ -537,7 +479,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/utils3/index"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #12", () => {
@@ -545,7 +487,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/utils4/index"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #13", () => {
@@ -553,7 +495,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/utils5/index"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #14", () => {
@@ -561,29 +503,24 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/timezones/pdt.mjs"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
-  it(
-    "invalid package target #15",
-    {
-      skip: "array fallback in exports field when first valid target file not found"
-    },
-    () => {
-      const result = resolver.sync(
-        fixture5,
-        "@exports-field/bad-specifier/non-existent.js"
-      );
-      assert.ok(result.error);
-    }
-  );
+  // skip: array fallback in exports field when first valid target file not found
+  it.skip("invalid package target #15", () => {
+    const result = resolver.sync(
+      fixture5,
+      "@exports-field/bad-specifier/non-existent.js"
+    );
+    expect(result.error).toBeTruthy();
+  });
 
   it("invalid package target #16", () => {
     const result = resolver.sync(
       fixture5,
       "@exports-field/bad-specifier/dep/multi1"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #17", () => {
@@ -591,7 +528,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/dep/multi2"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #18", () => {
@@ -599,7 +536,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/dep/multi4"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("invalid package target #19", () => {
@@ -607,7 +544,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/dep/multi5"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("should resolve the valid thing in array of export #1", () => {
@@ -615,7 +552,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/bad-specifier.js"
     );
-    assert.strictEqual(result.path, path.resolve(fixture5, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture5, "./a.js"));
   });
 
   it("should resolve the valid thing in array of export #2", () => {
@@ -623,7 +560,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/bad-specifier1.js"
     );
-    assert.strictEqual(result.path, path.resolve(fixture5, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture5, "./a.js"));
   });
 
   it("should resolve the valid thing in array of export #3", () => {
@@ -631,7 +568,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/dep/multi"
     );
-    assert.strictEqual(result.path, path.resolve(fixture5, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture5, "./a.js"));
   });
 
   it("should resolve the valid thing in array of export #4", () => {
@@ -639,7 +576,7 @@ describe("exportsFieldPlugin", () => {
       fixture5,
       "@exports-field/bad-specifier/dep/multi3"
     );
-    assert.strictEqual(result.path, path.resolve(fixture5, "./a.js"));
+    expect(result.path).toBe(path.resolve(fixture5, "./a.js"));
   });
 
   it("should not fall back to parent node_modules when exports field maps to a missing file (issue #399)", () => {
@@ -652,6 +589,6 @@ describe("exportsFieldPlugin", () => {
       path.resolve(fixture6, "workspace"),
       "pkg/src/index.js"
     );
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 });
