+++ napi/tests/extension-alias.test.mjs
@@ -1,12 +1,8 @@
-import { describe, it } from "node:test";
+import { describe, it, expect } from "@rstest/core";
 import { ResolverFactory } from "../index.js";
-import * as assert from "node:assert";
 import * as path from "node:path";
-import { fileURLToPath } from "url";
 
-const fixtureDir = fileURLToPath(
-  new URL("../../fixtures/enhanced_resolve/test/fixtures", import.meta.url)
-);
+const fixtureDir = path.resolve("fixtures/enhanced_resolve/test/fixtures");
 const fixture = path.resolve(fixtureDir, "extension-alias");
 
 describe("extension-alias", () => {
@@ -21,27 +17,27 @@ describe("extension-alias", () => {
 
   it("should alias fully specified file", () => {
     const result = resolver.sync(fixture, "./index.js");
-    assert.strictEqual(result.path, path.resolve(fixture, "index.ts"));
+    expect(result.path).toBe(path.resolve(fixture, "index.ts"));
   });
 
   it("should alias fully specified file when there are two alternatives", () => {
     const result = resolver.sync(fixture, "./dir/index.js");
-    assert.strictEqual(result.path, path.resolve(fixture, "dir", "index.ts"));
+    expect(result.path).toBe(path.resolve(fixture, "dir", "index.ts"));
   });
 
   it("should also allow the second alternative", () => {
     const result = resolver.sync(fixture, "./dir2/index.js");
-    assert.strictEqual(result.path, path.resolve(fixture, "dir2", "index.js"));
+    expect(result.path).toBe(path.resolve(fixture, "dir2", "index.js"));
   });
 
   it("should support alias option without an array", () => {
     const result = resolver.sync(fixture, "./dir2/index.mjs");
-    assert.strictEqual(result.path, path.resolve(fixture, "dir2", "index.mts"));
+    expect(result.path).toBe(path.resolve(fixture, "dir2", "index.mts"));
   });
 
   it("should not allow to fallback to the original extension or add extensions", () => {
     const result = resolver.sync(fixture, "./index.mjs");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   describe("should not apply extension alias to extensions or mainFiles field", () => {
@@ -55,18 +51,12 @@ describe("extension-alias", () => {
 
     it("directory", () => {
       const result = resolver2.sync(fixture, "./dir2");
-      assert.strictEqual(
-        result.path,
-        path.resolve(fixture, "dir2", "index.js")
-      );
+      expect(result.path).toBe(path.resolve(fixture, "dir2", "index.js"));
     });
 
     it("file", () => {
       const result = resolver2.sync(fixture, "./dir2/index");
-      assert.strictEqual(
-        result.path,
-        path.resolve(fixture, "dir2", "index.js")
-      );
+      expect(result.path).toBe(path.resolve(fixture, "dir2", "index.js"));
     });
   });
 });
