+++ napi/tests/incorrect-description-file.test.mjs
@@ -1,12 +1,8 @@
-import { describe, it } from "node:test";
+import { describe, it, expect } from "@rstest/core";
 import { ResolverFactory } from "../index.js";
-import * as assert from "node:assert";
 import * as path from "node:path";
-import { fileURLToPath } from "url";
 
-const fixtureDir = fileURLToPath(
-  new URL("../../fixtures/enhanced_resolve/test/fixtures", import.meta.url)
-);
+const fixtureDir = path.resolve("fixtures/enhanced_resolve/test/fixtures");
 const fixtures = path.join(fixtureDir, "incorrect-package");
 
 function p(...args) {
@@ -18,11 +14,11 @@ describe("incorrect description file", () => {
 
   it("should not resolve main in incorrect description file #1", () => {
     const result = resolver.sync(p("pack1"), ".");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 
   it("should not resolve main in incorrect description file #2", () => {
     const result = resolver.sync(p("pack2"), ".");
-    assert.ok(result.error);
+    expect(result.error).toBeTruthy();
   });
 });
